Updated per comments for assertion and tags
Also added few steps in feature to run it on Testng with scenarios.
Notes : Not able to remove import as using it in method and getting multiple errors.Also reviewed it with Soukaina too. Please let me know if you have any other suggestion for them we can go ahead with that.