Created new test @SLCR-381 WIP
New test will cover application results for when a user is SES limited in coco. This is a new test and it is still WIP.

I had to update the locator for the click back on the popup on application results page.

We found a css that could be used but the problem is when a user inputs an income between 12000 - 45000 it will show the verify your information new income MVR popup. So now it makes it be a list vs individual element. Now if someone reports an income higher than 45000 verify information popup does not appear so the weblist for the locator fails. The only way to get one locator was using the xpath. I used the xpath for both scenarios and it worked perfect.