Updates to Financial Agreements page in UI
SOL has made updates to the financial agreements page
The checkboxes have been updated to three boxes instead of one
The text on the page has ben updated.
This PR includes the steps for new checkbxes and steps to verify the text on financial agreements page.
New test case SLER-2206 which verifies the new functionality with multiple sinatures on the page
updated steps in SLER-533 with new checkboxes steps.